mqtt/handson/01-setup-and-basic-connection/requirements.txt
# Handson 01: Setup and Basic Connection - Python Dependencies
paho-mqtt>=1.6.1
rich>=13.7.0